+++ CHANGELOG.md
@@ -6,6 +6,8 @@ UNRELEASED
   https://gitlab.com/gitlab-org/gitaly/merge_requests/249
 - Implement RefService.FindAllBranches RPC
   https://gitlab.com/gitlab-org/gitaly/merge_requests/239
+- Implement FindAllTags RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/246
 
 v0.23.0